# maxmoc/api/openapi.MaxBotApi.yaml
openapi: 3.0.0
info:
  title: Max Bot API
  version: 0.0.32
  description: Бот-платформа мессенджера MAX. https://dev.max.ru/docs-api
...
paths:
  /answers:
    post:
      operationId: answerOnCallback
      summary: Ответ на callback
      description: |-
        Отправка ответа после того, как пользователь нажал на кнопку. Ответом
        может быть обновлённое сообщение. Не более двух ответов в секунду в один
        диалог, групповой чат или канал
      parameters:
        - name: callback_id
          in: query
          required: true
          description: Идентификатор кнопки, на которую нажал пользователь
          schema:
            type: string
            pattern: ^(?!\s*$).+
            minLength: 1
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleQueryResult'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Метод не разрешён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - messages
        - bot-sending
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallbackAnswer'
  /me:
    get:
      operationId: getMyInfo
      summary: Получение информации о боте
      description: Возвращает информацию о боте, который идентифицируется с помощью токена доступа `access_token`
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotInfo'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - bots
        - bot-sending
  /me/commands:
    patch:
      operationId: editMyCommands
      summary: Редактирование команд бота
      description: Добавляет, изменяет или удаляет команды бота. Для удаления команд передайте пустой массив `commands`
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotCommandsInfo'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - bots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotCommandsPatch'
  /messages:
    get:
      operationId: getMessages
      summary: Получение сообщений
      description: |-
        Возвращает информацию о сообщении или массив сообщений из чата. Нужно
        указать один из параметров — `chat_id` или `message_ids`
      parameters:
        - name: chat_id
          in: query
          required: false
          description: ID чата, чтобы получить сообщения из определённого чата. Обязательный параметр, если не указан `message_ids`
          schema:
            type: integer
            format: int64
          explode: false
        - name: message_ids
          in: query
          required: false
          description: |-
            Список ID сообщений, которые нужно получить (через запятую).
            Обязательный параметр, если не указан `chat_id`.

            ОТЛИЧИЕ ОТ ОРИГИНАЛА (осознанное, см. DEVIATIONS в scripts/compare.py):
            в оригинальной схеме этот параметр описан как массив без объявленного
            items вообще — недосмотр (тот же приём копипасты, что и у UserIdsList).
            Мы транскрибируем содержательный смысл: массив строк.
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: from
          in: query
          required: false
          description: Время, до которого будут запрошены все сообщения с начала чата (в формате Unix timestamp)
          schema:
            type: integer
            format: int64
          explode: false
        - name: to
          in: query
          required: false
          description: Время, начиная с которого будут запрошены все сообщения до конца чата (в формате Unix timestamp)
          schema:
            type: integer
            format: int64
          explode: false
        - name: count
          in: query
          required: false
          description: Максимальное количество сообщений в ответе
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageList'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'Ошибка доступа: нет прав на этот ресурс'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - messages
    post:
      operationId: sendMessage
      summary: Отправить сообщение
      description: |-
        Отправляет сообщение в диалог, групповой чат или канал. Можно отправлять
        не более двух сообщений в секунду в один диалог, групповой чат или канал —
        при превышении лимита ставьте сообщения в очередь или делайте задержку
      parameters:
        - name: user_id
          in: query
          required: false
          description: Если вы хотите отправить сообщение пользователю, укажите его ID
          schema:
            type: integer
            format: int64
          explode: false
        - name: chat_id
          in: query
          required: false
          description: Если сообщение отправляется в чат, укажите его ID
          schema:
            type: integer
            format: int64
          explode: false
        - name: disable_link_preview
          in: query
          required: false
          description: Если `true`, сервер не будет генерировать превью для ссылок в тексте сообщения
          schema:
            type: boolean
            default: false
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResult'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - messages
        - bot-sending
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMessageBody'
    put:
      operationId: editMessage
      summary: Редактировать сообщение
      description: |-
        Редактирует сообщения, отправленные ботом. В диалогах с ботом сообщения
        с кнопками `inline_keyboard` редактируются независимо от срока давности,
        остальные — если отправлены менее 7 суток назад; в групповых чатах и
        каналах — независимо от срока давности. Не более двух редактирований в
        секунду в одном диалоге, групповом чате или канале
      parameters:
        - name: message_id
          in: query
          required: true
          description: ID редактируемого сообщения
          schema:
            type: string
            minLength: 1
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleQueryResult'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - messages
        - bot-sending
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMessageBody'
    delete:
      operationId: deleteMessage
      summary: Удалить сообщение
      description: |-
        Удаляет сообщения в диалоге, групповом чате или канале. Бот должен быть
        администратором и иметь право на удаление сообщений. В диалоге бот может
        удалять только собственные сообщения, в групповом чате и канале — любые.
        Не более двух удалений в секунду в одном диалоге, групповом чате или канале
      parameters:
        - name: message_id
          in: query
          required: true
          description: ID удаляемого сообщения
          schema:
            type: string
            minLength: 1
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleQueryResult'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'Ошибка доступа: нет прав на этот ресурс'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - messages
        - bot-sending
  /messages/{messageId}:
    get:
      operationId: getMessageById
      summary: Получить сообщение
      description: Возвращает сообщение по его ID
      parameters:
        - name: messageId
          in: path
          required: true
          description: ID сообщения (`mid`), чтобы получить одно сообщение в чате
          schema:
            type: string
            pattern: '[a-zA-Z0-9_\-]+'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - messages
  /subscriptions:
    get:
      operationId: getSubscriptions
      summary: Получение всех подписок через Webhook
      description: Если ваш бот получает данные через Webhook, этот метод возвращает список всех подписок
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSubscriptionsResult'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - subscriptions
    post:
      operationId: subscribe
      summary: Подписка на обновления о новых событиях через Webhook
      description: Настраивает доставку событий бота через Webhook
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleQueryResult'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequestBody'
    delete:
      operationId: unsubscribe
      summary: Отписка от обновлений о новых событиях через Webhook
      description: Отписывает бота от получения обновлений о новых событиях через Webhook
      parameters:
        - name: url
          in: query
          required: true
          description: URL, который нужно удалить из подписок на WebHook
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleQueryResult'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - subscriptions
...
  /uploads:
    post:
      operationId: getUploadUrl
      summary: Загрузка медиафайлов
      description: |-
        Метод возвращает URL для загрузки медиафайла и токен для передачи
        загруженного файла во вложении к сообщению в чате или канале
      parameters:
        - name: type
          in: query
          required: true
          description: 'Тип загружаемого файла. Возможные значения: `"image"`, `"video"`, `"audio"`, `"file"`'
          schema:
            $ref: '#/components/schemas/UploadType'
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadEndpoint'
        '401':
          description: 'Ошибка авторизации: не предоставлен или недействителен access_token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - upload
        - bot-sending
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  schemas:
...
    BotCommandsInfo:
      type: object
      properties:
        commands:
          type: array
          items:
            $ref: '#/components/schemas/BotCommand'
          maxItems: 32
          description: Команды, которые поддерживает бот
      description: Информация о командах бота
...
    CallbackAnswer:
      type: object
      properties:
        message:
          type: object
          allOf:
            - $ref: '#/components/schemas/NewMessageBody'
          nullable: true
          description: Заполните это, если хотите изменить текущее сообщение
      description: Отправьте этот объект, когда ваш бот хочет отреагировать на нажатие кнопки
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        error:
          type: string
          description: Ошибка
        code:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Читаемое описание ошибки
      description: Сервер возвращает это, если возникло исключение при вашем запросе
...
    SimpleQueryResult:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: '`true`, если запрос был успешным, `false` — в противном случае'
        message:
          type: string
          description: Объяснительное сообщение, если результат не был успешным
      description: Простой ответ на запрос
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: access_token
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: https://platform-api2.max.ru
    variables: {}